// app/(default)/thing/item/types/GetSpecificItem.ts
export interface GetSpecificItem_getSpecificItem_assetAccount {
  __typename: "AssetAccount";
  _id: string;
  credit: GetSpecificItem_getSpecificItem_assetAccount_credit;
  paySchedule: GetSpecificItem_getSpecificItem_assetAccount_paySchedule[] | null;
}

export interface GetSpecificItem_getSpecificItem_itemFleet {
  __typename: "ItemFleet";
  _id: string;
  fleetName: string;
}

export interface GetSpecificItem_getSpecificItem_itemBatch {
  __typename: "ItemBatch";
  _id: string;
  deleteStatus: boolean | null;
  deleteAt: any | null;
  createdAt: any | null;
  updatedAt: any | null;
  type: ActorTypes;
  actionScope: ActionScope;
  actorName: string;
  profile: string | null;
  /**
   * batchNumber has a pattern of ProductClassBases+YYMM+123456
   */
  batchNumber: string;
  batchDate: any;
  /**
   * Please add description of batch!
   */
  description: string;
  batchState: BatchState;
  starting_code: string | null;
  /**
   * A key needed to generate the current token
   */
  secret_key: string | null;
  code_gen_type: CodeSystemType | null;
}

export interface GetSpecificItem_getSpecificItem_itemFirmware {
  __typename: "ItemFirmware";
  _id: string;
  /**
   * Version Numbers are controlled by OVES Production.
   */
  version: string;
  codeSystem: CodeSystemType;
}

export interface GetSpecificItem_getSpecificItem_codeGenerator {
  __typename: "CodeGenerator";
  _id: string;
  deleteStatus: boolean | null;
  deleteAt: any | null;
  createdAt: any | null;
  updatedAt: any | null;
  type: ActorTypes;
  actionScope: ActionScope;
  actorName: string;
  profile: string | null;
  hashRoot: string;
  hashTop: string;
  hashTopInitial: string | null;
  codeCount: number;
  hashIndex: number;
  codeReversalCount: number;
}

export interface GetSpecificItem_getSpecificItem {
  __typename: "Item";
  _id: string;
  deleteStatus: boolean | null;
  deleteAt: any | null;
  createdAt: any | null;
  updatedAt: any | null;
  type: ActorTypes;
  actionScope: ActionScope;
  actorName: string;
  profile: string | null;
  idType: ThingIDTypes;
  idString: string | null;
  description: string;
  creationDate: any;
  /**
   * Default ItemID is 'OVES'
   */
  oemID: string;
  /**
   * oemItemID cannot be null.  It must be provided in a conformal format PROD+YYMM+000000
   */
  oemItemID: string;
  /**
   * Identified a seller, such as 'ANGAZA'.
   */
  sellerID: string;
  /**
   * This is seller designated ID, unique within range.  Cannot be null.  Default same as oemItemID
   */
  sellerItemID: string;
  assetAccount: GetSpecificItem_getSpecificItem_assetAccount | null;
  itemFleet: GetSpecificItem_getSpecificItem_itemFleet | null;
  itemBatch: GetSpecificItem_getSpecificItem_itemBatch;
  itemFirmware: GetSpecificItem_getSpecificItem_itemFirmware;
  lifeCycle: string;
  codeGenerator: GetSpecificItem_getSpecificItem_codeGenerator | null;
}

export interface GetSpecificItem {
  getSpecificItem: GetSpecificItem_getSpecificItem;
}

export interface GetSpecificItemVariables {
  itemId: string;
}